Move contour wasm call into the web worker.
